fix: userRole & getUserRole error
