+++ app/spaces/[spaceid]/trustful/admin/page.tsx
@@ -8,7 +8,6 @@ import { Stack, Box } from '@mui/material';
 import { useCeramicContext } from '@/context/CeramicContext';
 import { Event, Space, SpaceEventData } from '@/types';
 import SubSidebar from '@/components/layout/Sidebar/SubSidebar';
-import { getUserRole } from '@/app/spaces/[spaceid]/trustful/service/backend/getUserRole';
 import { useTrustful } from '@/context/TrustfulContext';
 import { Address } from 'viem';
 import { AdminSection } from './components/AdminSection';
@@ -124,57 +123,6 @@ const TrustfulAdminPage = () => {
       return id?.[id?.length - 1];
     }
   }, [profile]);
-
-  //TODO: REFACTOR THIS WHEN API RETURNS ALSO LOOKING THE CONTRACT HASROLE
-  useEffect(() => {
-    const fetchUserRole = async () => {
-      if (address) {
-        await getUserRole(address as Address)
-          .then(async (data) => {
-            console.log('User role:', data);
-            if (data && data?.role) {
-              if (data.role == Role.NO_ROLE) {
-                const isRoot = await hasRole(ROLES.ROOT, address as Address);
-                const isManager = await hasRole(
-                  ROLES.MANAGER,
-                  address as Address,
-                );
-                const isVillager = await hasRole(
-                  ROLES.VILLAGER,
-                  address as Address,
-                );
-                console.log('isManagert', isManager);
-                console.log('isVillager', isVillager);
-                console.log('isRoot', isRoot);
-                if (isRoot) {
-                  setUserRole({
-                    address: address as Address,
-                    role: Role.ROOT,
-                  });
-                } else if (isManager) {
-                  setUserRole({
-                    address: address as Address,
-                    role: Role.MANAGER,
-                  });
-                } else if (isVillager) {
-                  setUserRole({
-                    address: address as Address,
-                    role: Role.VILLAGER,
-                  });
-                }
-              } else {
-                setUserRole({
-                  address: address as Address,
-                  role: data.role,
-                });
-              }
-            }
-          })
-          .catch((error) => console.error('Failed to fetch events:', error));
-      }
-    };
-    fetchUserRole();
-  }, [address]);
 
   return (
     <Stack direction="row" height="calc(100vh - 50px)" width="100%">
