+++ app/spaces/[spaceid]/trustful/page.tsx
@@ -7,7 +7,6 @@ import { Stack, Box } from '@mui/material';
 import { useCeramicContext } from '@/context/CeramicContext';
 import { Event, Space, SpaceEventData } from '@/types';
 import SubSidebar from '@/components/layout/Sidebar/SubSidebar';
-import { getUserRole } from '@/app/spaces/[spaceid]/trustful/service/backend/getUserRole';
 import { useTrustful } from '@/context/TrustfulContext';
 import { GiveBadge } from './components/GiveBadge';
 import { Address } from 'viem';
@@ -122,57 +121,6 @@ const TrustfulPage = () => {
       return id?.[id?.length - 1];
     }
   }, [profile]);
-
-  const getUserRoleByContract = async () => {
-    const isVillager = await hasRole(ROLES.VILLAGER, address as Address);
-    const isManager = await hasRole(ROLES.MANAGER, address as Address);
-    const isRoot = await hasRole(ROLES.ROOT, address as Address);
-
-    if (isVillager) {
-      setUserRole({
-        address: address as Address,
-        role: Role.VILLAGER,
-      });
-    } else if (isManager) {
-      setUserRole({
-        address: address as Address,
-        role: Role.MANAGER,
-      });
-    } else if (isRoot) {
-      setUserRole({
-        address: address as Address,
-        role: Role.ROOT,
-      });
-    }
-  };
-
-  useEffect(() => {
-    const fetchUserRole = async () => {
-      if (address) {
-        await getUserRole(address as Address)
-          .then(async (data) => {
-            console.log('User role:', data);
-            if (data && data?.role) {
-              //TODO: Check if the endpoint is returning correctly and if needed to check for Role.NO_ROLE ( double check confirmation here)
-              if (data.role == Role.NO_ROLE) {
-                await getUserRoleByContract();
-              } else {
-                alert('nao tem eh nada');
-                setUserRole({
-                  address: address as Address,
-                  role: data.role,
-                });
-              }
-            }
-          })
-          .catch(async (error) => {
-            console.error('Failed to fetch events:', error);
-            await getUserRoleByContract();
-          });
-      }
-    };
-    fetchUserRole();
-  }, [address]);
 
   return (
     <Stack direction="row" height="calc(100vh - 50px)" width="100%">
